I made final changes to the project

+++ BL/Bo/DroneInPackage.cs
@@ -12,7 +12,7 @@ namespace IBL
         public class DroneInPackage
         {
             public int ID { get; set; }
-            public int BatteryStatus { get; set; }
+            public double BatteryStatus { get; set; }
             public Location Location { get; set; }
             public List<ParcelInCustomer> DeliverieToCustomer = new ();
             public List<ParcelInCustomer> Delivery_list_customer_to_customer = new List<ParcelInCustomer>();
